requirements.txt
# =============================================================================
# Core Framework
# =============================================================================
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# =============================================================================
# AI/ML - LangChain Integration
# =============================================================================
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.10
groq==0.4.2

# =============================================================================
# AI/ML - Sentiment Analysis
# =============================================================================
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2

# =============================================================================
# Data Collection
# =============================================================================
praw==7.7.1
ccxt==4.1.92
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2
beautifulsoup4==4.12.2

# =============================================================================
# Data Processing
# =============================================================================
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2

# =============================================================================
# Storage
# =============================================================================
faiss-cpu==1.7.4
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.23

# =============================================================================
# Dashboard
# =============================================================================
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# =============================================================================
# Testing
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# =============================================================================
# Utilities
# =============================================================================
loguru==0.7.2
websockets==12.0
tenacity==8.2.3
structlog==23.2.0
